backend/requirements.txt
aioboto3==13.3.0
aiobotocore==2.16.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
attrs==24.3.0
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
charset-normalizer==3.4.1
dependency-injector==4.45.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
propcache==0.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
wrapt==1.17.0
yarl==1.18.3